server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
chatgptpy==1.0.8
colorama==0.4.6
cssselect2==0.7.0
Django==4.1.5
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
getenv==0.2.0
h11==0.14.0
idna==3.4
lxml==4.9.2
markdownify==0.11.6
multidict==6.0.4
openai==0.26.5
openpyxl==3.1.1
outcome==1.2.0
Pillow==9.4.0
pyChatGPT==0.4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tls-client==0.1.8
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.4.6
urllib3==1.26.14
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
